choosing page logic part1 done
