+++ resources/views/components/business-card.blade.php
@@ -1,5 +1,4 @@
 @props(['business'])
-{{-- !to check if it works  --}}
 <a href="/business/{{ $business->id }}/products">
     <article>
         <h3>{{ $business->name }} </h3>
@@ -7,7 +6,6 @@
             <p>works</p>
         @endif
         <div>
-            {{-- !to fix --}}
             <img src="/storage/businessImages/restaurantGeneral.png" alt="restaurant">
         </div>
     </article>
