+++ routes/Routes/businessRoutes.php
@@ -16,5 +16,3 @@ Route::get("/business/{id}/edit", [BusinessController::class, "edit"])->where("i
 Route::put("/business/{id}", [BusinessController::class, "update"])->where("id","[0-9]+");
 
 Route::delete("/business/{id}", [BusinessController::class, "destroy"])->where("id","[0-9]+");
-
-Route::post("/choosing", [BusinessController::class, "getBy2kmRadius"]);
