+++ routes/Routes/homePageRoutes.php
@@ -9,6 +9,9 @@ Route::get("/", [HomePageController::class, "home"]);
 //Route for the choosing page
 Route::get("/choosing", [HomePageController::class, "choosing"]);
 
+//Route to trigger choosing logic
+Route::post("/choosing",[HomePageController::class, "getBy2kmRadius"]);
+
 Route::get("/choosing/sell", [HomePageController::class, "sellPath"]);
